refactor(context): extract async event notification into new class
